refactor endpoint trace rule nodes
